requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
urllib3==2.2.3
python-dotenv==1.0.1
tenacity==9.0.0
pandas==2.2.2
openpyxl==3.1.5
psycopg2-binary==2.9.9
